pyproject.toml
[tool.poetry]
name = "recipe-app"
version = "0.1.0"
description = "Recipe API for Udemy Course"
authors = ["Darren Weber <dweber.consulting@gmail.com>"]
license = "Apache 2.0"
readme = "README.md"

packages = [
    {include = "employees"},
    {include = "recipe_app"},
    {include = "recipe_api"},
]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.0"
django-ninja = "^0.21.0"
djangorestframework = "^3.14.0"
python-decouple = "^3.8"
dj-database-url = "^1.2.0"

[tool.poetry.group.dev.dependencies]
black = "^23.0"
isort = "^5.0"
pre-commit = "^3.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.0"

[tool.poetry.group.docs.dependencies]
mkdocs = "*"

[tool.isort]
profile = "black"
force_single_line = true
extend_skip_glob = ["**/migrations/*.py", ]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
